Move function to utils. Change to import error handling.
